Fix Integer warnings

+++ FrameLib_Objects/Routing/FrameLib_Select.cpp
@@ -38,7 +38,7 @@ std::string FrameLib_Select::objectInfo(bool verbose)
 
 std::string FrameLib_Select::inputInfo(unsigned long idx, bool verbose)
 {
-    if (idx == mNumIns)
+    if (idx == static_cast<unsigned long>(mNumIns))
         return parameterInputInfo(verbose);
     else
         return formatInfo("Input #", idx);
